[change] Separated translation class
